+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
 ## v0.0.1 [Unreleased]
+- 16/05/18 feat(web): add web server, middleware and router and get implementation without range (SFR)
+           refact(dao): fix panic on wrong object id format
+           chores(vendor): upate vendors with web dependency
 - 16/05/11 feat(dao): add model of spirits, mongo dao, factory and update vendors (SFR)
 - 16/05/10 feat(all): project start, add command line, vendors, makefile, docker, logger, tests and statistics (SFR)